refactor(ui): restructure actuator components
- Deleted the actuators/ui.tsx file and split its components into separate files for better modularity.
- Moved ProportionalSlider to actuators/ui/proportionalSlider.
- Moved PumpSwitch and SolenoidSwitch to actuators/ui/actuatorSwitches.
- Updated imports in page.tsx to reflect the changes in component locations.

issue #24
